refactor: routes.js 파일 명을 App.js로 재변경
