+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './routes';
+import App from './App';
 import GlobalStyle from './Styles/GrobalStyle';
 
 ReactDOM.render(
